Implement TargetObject Spawning position and Bug Fix for Button Clicks
Added functionality for spawning TargetObjects at the position where the mouse is pressed. Continuous spawning is now possible by holding the mouse button, with a set delay of 0.25 seconds to prevent macro abuse.

Fixed a bug where buttons failed to respond to user clicks and instead initiated TargetObject spawning. The issue was resolved by adding the spawn objects function to the buttons' inspector within the engine UI.

Improved the game's input handling to accurately detect whether the input is a mouse click or a touch event on a smartphone, and to execute the appropriate action: spawning a TargetObject or clicking a button.