Implemented Scriptable Objects for Food-Types and Integrated Cost Management System
- Day 16:
16.1.
Developed the Scriptable-Object script to manage various food-related data within the game.
16.2.
Created 5 distinct Scriptable-Objects with predefined values for later use, dividing them between 3 for the Free version and 2 for the Paid version of the game.

- Day 17:
17.1.
In the Game-Manager script, introduced a list of type “Food-Properties” to store all Scriptable-Objects, representing different food types.
17.2.
Each newly initiated TargetObject is now assigned a current Food-Type from this list.
17.3.
Implemented a cost management system where the In-Scene-Money is decreased by the assigned Food-Type's cost whenever a TargetObject is initiated.
17.4.
Added a restriction to ensure that players can only initiate a new TargetObject if their In-Scene-Money is greater than the cost of the assigned Food-Type.
17.5.
Enhanced the UI by adding a new button in the UIManager script that allows players to upgrade the Food-Type assigned to the next TargetObject, with this interaction managed via the Game-Events system.