Added Gumbo The Angler (GTA) and Improved Protective Pet Behavior and Movement Strategies
Day 55:
55.1.
I introduced a new pet named “Gumbo The Angler (GTA)”:
55.1.1.
When an enemy object is found in the scene, the GTA pet uses the “Bounds And Positioning Manager” script to find the farthest corner on the screen from the enemy object and move towards it.
55.2.
I modified both scripts, the “Threatened Swimming State” and “Game Manager” scripts:
55.2.1.
I added a new static function to the “Game Manager” script that identifies the best protective pet from the existing protective pets in the scene and then returns an Enum key.
55.2.2.
I modified the “Determine Movement Strategy” function to select only the appropriate movement strategy instead of identifying the type of protective objects and setting the movement strategy.
55.3.
I modified the “Single Target Following Movement Strategy” and added a “Random Dispersion Factor” to disperse the movement direction around the target object.