Enhanced Game Performance, Hunger Indicators, and Protective Pet System Refinements
Day 56:
56.1.
I fixed some bugs related to the system scripts which increased the performance of the game 2 times.
56.1.1.
I redesigned the search mechanism to be state-based upon invocation.
56.2.
I added a new feature that presents the hunger situation for Main-Fish objects.
56.3.
I refactored the protective pet identification functionality which is now decided inside the GameManager script itself.
56.4.
I introduced a new pet object that when instantiated enables the hunger arrow over the fishes (or objects) that can feel hungry.
56.4.1.
Another feature of this object will be added later, which is detecting where the next enemy object will be instantiated.
56.5.
I moved the protection delay data from WTW pet’s state to the data level script, allowing it to be set by the level itself. In other words, modifying the difficulty of the level by levels data.